Rename SyncingFileSystem to RaptorLocalFileSystem
